Add files via upload
From a Tier 1 IT position I wanted to find an efficient way to display how much free space domain PCs had available on their drive upstream of files deployment. This is where this script comes handy as it lists the latter across the network rather than checking it manually on each machines.